+++ src/widgets/input/lugia.input.zh-CN.json
@@ -71,7 +71,7 @@
       "defaultValue": false
     },
     "type": {
-      "type": "string",
+      "type": "InputType",
       "desc": "input类型",
       "defaultValue": "text"
     },
@@ -175,6 +175,7 @@
   },
   "type": {
     "InputSize": ["small", "default", "large"],
+    "InputType": [ "text","password"],
     "ValidateStatus": ["success", "error"],
     "InputValidateType": ["top", "bottom", "inner", "default"],
     "ChangeType": {
