+++ src/widgets/input/lugia.textarea.zh-CN.json
@@ -1,0 +1,176 @@
+{
+  "widgetName": "Input",
+  "title": "段落文本输入框",
+  "desc": "常用于昵称,名称,表格内容等填写.",
+  "props": {
+    "disabled": {
+      "type": "boolean",
+      "desc": "禁用状态,是否不可用",
+      "defaultValue": false
+    },
+    "placeholder": {
+      "type": "string",
+      "desc": "段落文本输入框输入提示信息"
+    },
+    "defaultValue": {
+      "type": "string",
+      "desc": "默认显示内容"
+    },
+    "value": {
+      "type": "string",
+      "desc": "显示内容"
+    },
+    "autoFocus": {
+      "type": "boolean",
+      "desc": "是否自动获取焦点",
+      "defaultValue": false
+    }
+  },
+  "events": {
+    "onChange": {
+      "desc": "内容改变时触发",
+      "args": [
+        {
+          "name": "event",
+          "desc": "改变内容的DOM事件",
+          "type": "ChangeType"
+        }
+      ]
+    },
+    "onFocus": {
+      "desc": "段落文本输入框获得焦点时触发",
+      "args": [
+        {
+          "name": "event",
+          "desc": "获取焦点的DOM事件",
+          "type": "FocusEvent"
+        }
+      ]
+    },
+    "onBlur": {
+      "desc": "段落文本输入框失去焦点时触发",
+      "args": [
+        {
+          "name": "event",
+          "desc": "失去焦点的DOM事件",
+          "type": "FocusEvent"
+        }
+      ]
+    },
+    "onKeyUp": {
+      "desc": "键盘按下去并松开后执行",
+      "args": [
+        {
+          "name": "event",
+          "desc": "改变键盘输入的内容",
+          "type": "KeyboardEvent"
+        }
+      ]
+    },
+    "onKeyDown": {
+      "desc": "键盘按下任何键时触发",
+      "args": [
+        {
+          "name": "event",
+          "desc": "改变键盘输入的内容",
+          "type": "KeyboardEvent"
+        }
+      ]
+    },
+    "onKeyPress": {
+      "desc": "键盘按下任何键时触发",
+      "args": [
+        {
+          "name": "event",
+          "desc": "改变键盘输入的内容",
+          "type": "KeyboardEvent"
+        }
+      ]
+    },
+    "onClear": {
+      "desc": "清除段落文本输入框内容时触发",
+      "args": [
+        {
+          "name": "event",
+          "desc": "清除段落文本输入框内容事件",
+          "type": "Object"
+        }
+      ]
+    },
+    "onEnter": {
+      "desc": "当键入回车时触发事件",
+      "args": [
+        {
+          "name": "event",
+          "desc": "当键入回车时触发",
+          "type": "KeyboardEvent"
+        }
+      ]
+    }
+  },
+  "type": {
+    "ChangeType": {
+      "newValue": "string",
+      "oldValue": "string",
+      "event": "SyntheticEvent"
+    }
+  },
+  "category": ["数据录入"],
+  "needExport": true,
+  "theme": {
+    "Container": {
+      "name": "段落文本输入框外部容器",
+      "desc": "段落文本输入框外部容器",
+      "normal": [["width"], ["height"], ["margin"]]
+    },
+    "Textarea": {
+      "name": "段落文本输入框主体",
+      "desc": "段落文本输入框主体结构",
+      "normal": [
+        ["width"],
+        ["height"],
+        ["fontSize"],
+        ["font"],
+        ["color"],
+        ["background"],
+        ["border"],
+        ["borderRadius"],
+        ["cursor"],
+        ["opacity"]
+      ],
+      "hover": [
+        ["border"],
+        ["borderRadius"],
+        ["cursor"],
+        ["background"],
+        ["opacity"],
+        ["boxShadow"]
+      ],
+      "active": [["boxShadow"], ["border"], ["borderRadius"], ["cursor"], ["background"]],
+      "disabled": [
+        ["fontSize"],
+        ["font"],
+        ["color"],
+        ["background"],
+        ["border"],
+        ["borderRadius"],
+        ["cursor"],
+        ["padding"],
+        ["opacity"]
+      ]
+    },
+    "Placeholder": {
+      "name": "段落文本输入框提示信息文字",
+      "desc": "段落文本输入框提示信息文字",
+      "normal": [["color"], ["fontSize"], ["font"]]
+    },
+    "InputClearButton": {
+      "name": "段落文本输入框清除图标",
+      "desc": "段落文本输入框后缀清除图标",
+      "normal": [["color"], ["fontSize"]],
+      "hover": [],
+      "clicked": [],
+      "disabled": []
+    }
+  }
+}
